Implemented the request for completion in Elisp.
The exploratory study was implemented in Emacs Lisp, and is ready to be
integrated with Company mode to provide the candidates for completion at
point.

Currently, however, it is necessary to evaluate (M-S-:)
(company-godot-gdscript-process-request-completion-at-point) in a
GDScript file to request the possible completions, as the integration
with Company mode has yet to be written.